Initial Load of Visual Studio Solution
Please be kind in your comments. This is my first GitHub project upload
and is being placed into Open Source by request. I removed several
ReSharper and temporary files. Hopefully I haven't deleted any important
ones.  The references to Design are related to the Expression Design
product.

In 2008 I reverse-engineered to the best of my ability the Adobe
proprietary ASE file format and documented it here:
http://iamacamera.org/default.aspx?section=develop/microsoft&id=109

This project is the driver behind this online converter site:
http://iamacamera.org/sandbox/aseconvert/

I'm releasing this as free open source code MIT if you need a name. Let
me know if there is a more official way to declare this free for all and
take it as is.